monitor/pyproject.toml
[project]
name = "happy-vibe-monitor"
version = "0.1.0"
description = "Happy Vibe 桌面监控器 - 系统托盘应用"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pystray>=0.19.5",
    "Pillow>=10.0.0",
    "plyer>=2.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "respx>=0.21.0",
]

[project.scripts]
happy-vibe-monitor = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
